requirements.txt
torch==1.12.1
torchvision==0.13.1
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib >=3.4.0
seaborn >=0.11.0
pillow >=8.0.0
psutil >=5.9.8	